Refactor to add HR and Accounting namespaces
