+++ BethanysPieShopHRM/Accounting/Account.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace BethanysPieShopHRM.Accounting
+{
+    public class Account
+    {
+
+    }
+}
