+++ BethanysPieShopHRM/BethanysPieShopHRM.csproj
@@ -11,4 +11,8 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="HR/" />
+  </ItemGroup>
+
 </Project>
