+++ BethanysPieShopHRM/HR/EmployeeType.cs
@@ -1,6 +1,5 @@
-namespace BethanysPieShopHRM
+namespace BethanysPieShopHRM.HR
 {
-
     enum EmployeeType
     {
         Sales,
