+++ BethanysPieShopHRM/HR/WorkTask.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace BethanysPieShopHRM
+namespace BethanysPieShopHRM.HR
 {
     internal struct WorkTask
     {
